LBT - Lua based tables - initial commit
The bones of the project are here. The organisation of files should be
complete: lbt.sty for the Latex import and lbt-[012345]-*.lua for the
Lua code.

Some implementation is done, but not much. Just enough to check that the
file organisation is sensible and to have something to document. There
is probably not even enough implementation to run anything at the
moment.

The main thing has been to get a bit of documentation in place that
shows some code to _outline_ the implementation.

I am about to start actual implementation, so this is a good place to
make a first commit.