Set up several files for Basic and "contrib"
I was going to have a "builtin" Basic template (always loaded) and some
"contrib" templates like contrib.Questions and contrib.WS0, which would
be loaded on demand.

But the design/implementation was starting to get a bit funky and the
prefix "contrib" is a bit long, so I have changed and simplified things.

There is now just lbt.Basic and lbt.Questions and lbt.WS0 and ..., each
of which lives in its own file lbt-7-lbt-Questions.lua etc. All of them
are fully pre-loaded, which is (very) slightly wasteful of memory, but
the conceptual simplification is worth it. The template _names_ are
lbt.*, which means they won't clash with user templates, they clearly
denote their origin, and lbt is short.

lbt.Basic will always be available; no explicit sourcing needed. It is a
special case in the code.

While I am writing at length, I keep thinking I would like a bit of
vocab for the two different types of template that there are in
practice:

 * a "Structure" template like WS0 does page layout and tends not to
   have any of its own functions. It definitely implements _expand_.

 * a "Content" template like Questions does not do page layout, rather
   it provides functions like Q, MC, ANSWER, etc. It does not implement
   _expand_, but uses the default expand in case someone wants to use it
   as the actual template.

I think those names are good. They will help clarify documentation and
may even end up being used in code, to streamline template registration.

NOTE All I have done in this commit is introduce some new files
(skeleton contents at the moment) and renamed one or two others. All the
files are loaded in lbt.sty. But there is (re-)work to be done in terms
of registering and resolving the templates, and (especially) adding user
templates to the registry.

But the work coming up will be easier than it was going to be before
this change.