Success! T_expand_Basic_template_1 passes
This involves expanding two TEXTS (one with a vspace argument) and an
ITEMIZE (including an option argument [topsep=0pt]).

Getting this to work involved a number of improvements to the legacy
code. When a token function is called, it is now passed nargs and args.
It returns one of:
 * 'ok', latex
 * 'nargs', spec      to signal an invalid number of arguments
 * 'error', msg       to indicate an error occurred

For tokens that take an optional "option" argument, the function
lbt.util.extract_option_argument(args) is very useful.

There is a lot of work now to update the rest of Basic to this style,
but it will be worth it.