sources and styles are now global variables
lbt.const.{sources,styles}

I don't love this and have an idea to improve it.

Passing sources around as a parameter is OK and I might return to it, or
I might make a HOF resolver(sources) --> function(token) which does
token resolution with a memo cache.

But passing styles as a parameter is unwieldy because it would have to
go to every template function like Questions.Q or Letter.SIGNOFF etc.
This is not good. Using a global variable here seems the only way. The
template functions can then call lbt.api.style('Q.vspace') etc.

Sources and styles are different concepts but they are kind of pairs in
that they have built-in and discretionary aspects. And they are both
needed to turn tokens into Latex. So it has always seemed that they
would need to be passed around together, and I was loath to treat access
to them differently. Hence I have made them _both_ global variables in
this commit but will soon change that.