Started to implement argument specifications
I have put the specs in the Basic template, like
    a.NEWCOMMAND = 3
    a.ITEMIZE = '1+'

I have implemented the code that parses these and makes a spec like
    { spec = '3+', min = 3, max = 9999 }

This 'spec' is now available in the template object through
    td.arguments['ITEMIZE']

I have _not_ implemented the code that checks the number of arguments on
expansion. That is the next commit.