Progress towards implementing styles
I have gone back and forth on how to approach this. I was intending to
use a function in token implementations, like lbt.api.style('Q.color').
But I didn't like the global data that would require, and I like the way
token resolution works, so ultimately I decided to use that approach.

There is significant documentation above the function lbt.api.style,
even though that function is no longer part of the design! It remains
there for now so I might put it elsewhere for future documentation.

The implementation is incomplete. Where there used to be 'resolver', the
word is now 'token_resolver' so we can also have 'style_resolver', or tr
and sr for short. lbt.fn.style_resolver is mostly implemented and is
nice. But it doesn't yet look in document-wide or template-local style
modifications. It only looks in the templates in the source list
provided.

To complete the implementation, I need to:

 * implement \lbtStyle{...} which calls lbt.api.add_style_settings or
   something

 * modify template registration to create collapsed style dictionaries