Registers implemented - WIP
Progress is pretty good. I have written a good unit test. Most aspects
of it are working (assignment, expansion, expiry). It seems like the
only thing not working (and no proper attempt yet to implement) is
mathmode. I am not 100% clear on the desired behaviour so I'll have to
work that out.

So, current commit: test not passing, but good progress nonetheless.