Overhaul of logging and debugging
There is no more 'debug' file, or function lbt.dbg.

lbt.log now takes a compulsory "channel". It can be a number 0-4 or a
string for a thematic channel.

0 = ANN, 1 = ERROR, 2 = WARN, 3 = INFO, 4 = TRACE

Existing thematic channels are:
 * read
 * parse
 * emit
 * register
 * templates

The author selects the channels they want thus:
  \lbtLogChannels{1,2,3,emit}
  \lbtLogChannels{all}

There is currently no way to add or remove a channel. This might be
revisited, because "all except trace" could be desirable.

Coming up: the !DEBUG pragma sets all channels for the duration of that
expansion, then resets the channels as they were afterwards.