Progress on lbt.WS0 and lbt.Questions
I am creating a document using lbt in a separate directory (not part of
this repository). This is spurring implementation and improvements to
WS0 and Questions, which were taken in the first instance from my old
GSC code. I am getting somewhere but the output has problems that I need
to diagnose. Hence it is time for a progress commit.

In this commit:
 * big changes to lbt.WS0 and lbt.Questions, obviously
 * lbt.util.latex_expand_content_list
 * lbt.util.content_meta
      - both of these are designed to be used within template code,
        and there will be more such things
 * fixed bug if argument spec not provided
 * made various things optional in template spec: init, expand,
   arguments, styles
      - this might make some unit tests fail, since Basic won't have an
        expansion function -- fix soon if necessary
 * lbt.fn.impl.latex_message_token_not_resolved - stylistic bug fixed
 * a bit of log file improvement