Ironed out the problem plaguing \integral
I needed to ensure the argument to the Lua function was not expanded by
Tex first. That is, we want to receive '\frac{...}{...}' and not the
lower-level Tex code that this expands to.

All it took was \unexpanded{#1}, but I didn't know about that and it
took a long time to find out.