lbt.Math macro <simplemath> implemented
This allows Latex code like
  $$ \sm{forall n in \Real, cos2n > 0} $$

The output of that macro would be
  \ensuremath{\forall n \in \Real, \cos^{2} n > 0}

Notice the addition of backslashes to forall and in, and the
sophisticated parsing of cos2n. Also, 'csc3th' -> '\csc^{3} \theta'.

The extent of its functionality is hinted by these lines. More will be
added as the need is encountered.

  local trig = makeset'sin cos tan sec csc cot'
  local other = makeset[[equiv forall exists implies to in]]
  local alpha = makeset[[alpha beta gamma delta epsilon theta pi phi omega]]
  local abbrev = makemap[[al alpha be beta ga gamma de delta th theta]]