Experimenting with improving simplemath
Little code here. Just starting to experiment with more sophisticated
patterns. Not getting very far, and I need to move on with other things.

The aim is to be able to parse something like
  lim_{n to infty} 1/n = 0 \text{ and } forall x in \Real, x^2 \ge 0
into
  \lim_{n \to \infty} 1/n = 0 \text{ and } \forall x in \Real, x^2 \ge 0

[aside: I'd like to turn 1/n into \frac{1}{n} but that's another story]

The problem is I split the input by space and operate on each space-
separated word, and 'lim' and 'infty' are not whole words.

* I want to pick up \text{...} and leave it alone, as no expansions
  should take place in there.
* I want to leave \Real alone - already achieved, but keep in mind
* I want to pick up lim and infty as words

It is challenging. pl.sip is fairly good but only gets so far. lpeg
might be able to help, but I don't know how to use that yet. I foresee a
lot of custom code being necessary.