(Re)implementation of QQ* and MC*
It is very nice. I no longer use tables or enumerate to produce the
sub-qestions. I use parbox to space things out appropriately (both to
create a label width and a question width). Also I allow the use of
hfill instead to create even spacing rather than equal columns.

Examples:
  QQ* [ncols=3, vspace=2em] :: text1 :: text2 :: text3 :: text4 :: ...
  QQ* [hpack=hfill] :: text1 :: text2 :: text3

The QQ* can be replaced with MC* and the correct result will ensue. QQ*
and MC* are styled differently (indentation, colour, other spacing), but
are typeset by the same underlying code.

(It is a probably future improvement to allow those spacing and colour
settings to be set as a style.)

Other changes in this commit:
* Implement E325_invalid_return_from_template_function(token, result)