Progress commit on branch lpeg-parsing
Same as last commit (a single unit test passes), except now the
ParsedContent class lives in fn instead of in parser. The demarcation is
clearer. lbt.parser does _initial_ parsing only, with no pragmas and no
indexing. So the function is called lbt.parser.parsed_content_0 and the
thing it returns is known as pc0. Then lbt.fn.parsed_content grabs this
and constructs the ParsedContent object with pc0 and pragmas. The
ParsedContent class handles index creation.

This commit continues the task of removing old code, and has flagged
some more for removal.