Progress commit on branch lpeg-parsing
Good progress. I just ran the code for the first time on real documents.
The code does not yet attempt to emit real Latex, but it does parse all the
documents and emit a placeholder with the eid and the title. Doing this
forced me to update code in several places and fix some deficiencies, so
that is excellent.

Some of the changes:
 * lbt.log now only calls string.format if there are enough arguments
   for it. There was a bug where the message contained % and so the
   formatter expected an argument but got none.
 * Some lbt.fn.pc.* function calls replaced with the object-oriented
   pc:*
 * Old-style content containing @META instead of [@META] is
   automatically updated, so I can test this code on real documents
   without having to update them in that way.
 * Comment lines stripped.