Progress commit on branch lpeg-parsing
Incomplete progress, but I'm getting closer to actually emitting some
Latex.

* pc methods template_object_or_error, extra_sources, extra_styles
  - note that 'styles' will become 'options' soon enough
* generally changed old-style pc code to new-style

Now we're up to parsed_content_to_latex_single, and the focus shifts to
the new format of parsed content. No longer is it line by line. It is
command by command, for what it's worth. And the new setup with optargs
and kwargs and posargs is not compatible with the existing code. Fun
times ahead.