Big updates to get vector.format working!
Renaming of variables and functions:
 - lbt.system.document_wide_styles -> document_wide_options
 - lbt.api.add_styles -> add_document_wide_options
 - \lbtStyles -> \lbtDocumentWideOptions

lbt.util.get_option_for_macro introduced, to make it clear that this
util function is only for use within macros like 'vector' or 'integral',
etc. This function gives clear error messages.

ParsedContent:local_options() now exits the program if STYLES is used or
if OPTIONS cannot be parsed to a dictionary.

OptionLookup:_lookup() performs a lookup that returns nil rather than
raising an error if the option cannot be resolved. This is so that
lbt.util.get_option_for_macro can raise its own error message, giving
clearer information to the user.

Option lookup now works if an option is set to false. We need to
distinguish between false and nil. That was a subtle bug.

OptionLookup:__tostring fixed and improved.

lbt.fn.style_string_to_map removed.

When an option lookup fails, the OptionLookup object is printed to give
more information.

Added errors E945_invalid_option_dictionary_wide and
E946_invalid_option_dictionary_narrow in case the user writes text that
cannot be parsed as a dictionary.

Various small updates to lbt.Math.

lbt.Article updated (but not properly tested yet).

lbt.Basic: starting an experiment with ITEMIZE.sep. It hasn't gone
anywhere yet, but I think we need a convenient way of specifying how
tight or loose we want a list to be. Like a value of 1 could mean 0.1em,
plus or minus 0.05em. Something like that.

lbt.util.get_style slated for removal