ALIGN(*) spreadlines option fixed
More generally: if you have an option that should default to nil, just
be aware that it won't parse to actual nil but to the string 'nil'. That
is fine; just test for that.

(We can't parse it to nil because then it won't appear in the options
dictionary and the option lookup will fail.)