Corrected QQ.vspace option
I changed o('QQ.vspace') to o.vspace. Although I'm pretty sure
o('vspace') would work too. Not that it's better or anything.

Change afoot. 'vspace' subtly communicates space _before_ the item,
whereas in this case we are setting vertical space _after_ the question.
I will change it to 'room' (as in, room to answer the question).

I think I will standardise on 'vsp' for vspace, so it takes up less
attention in the document code.