Article: finer control over parindent and parskip
Removed option Article.parstyle, and now set options parskip and
parindent directly. Default is 15pt indent (same as Latex) and a little
bit of parskip. I like that style in writing I have recently seen. Now I
am ironing out wrinkles in other parts of the code (ALIGN, ITEMIZE, ...)
to have consistent control over whether a command constitutes a
paragraph or not.

It is all well to have TEXT and TEXT*, but I am tiring of ITEMIZE and
ITEMIZE*, and ALIGN and ALIGN* etc. It causes duplication of opargs. A
better approach would be a 'nopar' oparg.

With ALIGN, I think I want to generalise it to MATH or EQUATIONS or
something, supporting opargs {align, gather, multiline, ...}. There is
no need for those to all be their own command. Then equation numbering
logic would be centralised.

I could generalise ITEMIZE and ENUMERATE to LIST, but will hasten slowly
there. The thought would be LIST with opargs {item, enum, ascii} with
the latter using the asciilist package to support nested lists easily.

But yeah, using * to suppress paragraphs has got to stop. Oparg 'nopar'
is the way to go.