Article template: cosmetic changes
HEADING now uses \section*, and that is it. Use a Latex command (or
three) to change the style. (In my Article documents so far, I use the
memoir class, which allows me to do \setsecheadstyle{\normalsize\scshape\bfseries}).

ABSTRACT now enforces no paragraph indentation and a small parskip.
Having indents in the abstract looks ugly. (It's probably the case that
an abstract should only be one paragraph...)