Oparg .vspace changed to .prespace throughout
vspace can be ambiguous: where does this space get inserted? "Of course"
it is before the item, but with QQ* is is after. So calling it prespace
makes that clear.

I intend for prespace and nopar to be universal options that
automatically apply to every command, and be handled in one place. But I
need to think about how to achieve that.

So for now it is just a change of name.