Fixed failing tests
The way we now handle \par in TEXT (outside the implementation of TEXT)
changes the return value of latex_for_commands: it is a list, and \par
appears as its own element rather than part of the string before it. So
I had to change the expected value of several tests.

Also in this commit:
 * the 'nopar' logic in fn:latex_for_command was reversed. We check for
   'par' == true instead of 'nopar' == false. Note: I have not yet
   embarked on implementing noX opargs.
 * TestQuestions.lua updated to return vspace on its own line, just like
   the real Questions.lua.