Option noX implemented (in a basic manner)
You can now write '.o nopar' instead of '.o par = false'. You can only
write it in a local oparg, so:
  TEXT .o nopar :: Hello                   [yes]
  OPTIONS TEXT.nopar                        [no]

It is resolved at Latex-emission time. Looking up option 'par' will
check whether 'nopar' is specified locally before going on to check the
document-wide and template settings.

Specifying 'nopar' is reasonable because 'par' is a thing and is
boolean. However, doing something like
  Q .o nobeer  :: What is the capital of Mauritius?
  Q .o nocolor :: What is the capital of Bolivia?
should be an error. In the first line, 'beer' is not an option and so
neither is 'nobeer'. In the second line, 'color' is an option but not a
boolean one, so 'nocolor' does not make sense.

There _should_ be a check to detect the problems above, but I am not
implementing it at this stage:
  1. Doing so would be complex. The interaction between local opargs,
     the cache, and template settings is settled and I don't want to
     unsettle it.
  2. This 'noX' feature is likely to be used for 'nopar' and little
     else.