New template lbt.PhotoGallery
It allows you to write code like the following.

    PHOTOGALLERY
     :: (folder) images-wedding
     :: (per_row) 2
     :: (max_height) 7cm
     :: (feature) 2..5 27

The result being a grid of photos, several per page, all much the same
size, with a number under each one. The purpose is to aid selection from
a large collection.

The 'feature' causes certain photos to appear as larger floats. They are
feature photos.

Extra functionality to be supported soon:
 - (include) to say which numbers should be included
 - (exclude) to exclude some
 - (mark) to subtly select some photos as "marked"

Also in this commit:
 - lbt.util.parse_numbers_and_ranges(text)
   Input: '2 5 6..8 19 25..28'    Output: { 2,5,6,7,8,19,25,26,27,28 }