Fixed bugs in simplemath
In particular, \sm{a \text{ is greater than } b} now parses correctly.
It was producing a \\text instead of \text, because 'text' is a special
word that expands to '\text'. Now we look especially for commands before
we look for general text.

This took a bit of fiddling, and I had to ensure \, and \! are picked up
as commands, but here we are.