(WIP) Exploring more bugs, need to look closely at template rego
If `s` is a source (e.g. lbt.Basic), then `s.default_options` is a map
like
  CHAPTER.starred   -> false
  ENUMERATE.compact -> false
  TABLE.position    -> htbp
  ...

This is good for looking up a qualified key, which is all we have done
so far. It is not good for making an "opargs spec" for (say) TABLE, so
that a TABLE invocation can have its opargs checked before expansion.
(We have not tried to do that before; we are trying to do it now.)

So a "better" value for `s.default_options` (which, by the way, will
soon be renamed to `s.opargs`) could be:
  CHAPTER -> { starred = false }
  TABLE   -> { float = false, position = htbp }
  ...

This would make it easy for me to build an oparg spec. It would make it
harder to just look up a qualified key to obtain a value. A solution to
_that_ problem could be to change the way I use qualified keys. That
problem is really limited to OptionLookup and util.get_option_for_macro.
It can be solved there, by using { 'TABLE', 'position' } as a qualified
key for lookup, while still using 'TABLE.position' for caching.

I think I will proceed with that idea.

But first, I'm committing the changes I've made. Mostly a tuning of the
code in ExpansionContext:resolve_oparg. The way it handles starred
opcodes is now neater. Also other minor code improvments, and a new test
or two.