(WIP) Working towards implementing new oparg spec style
See previous commit (17a441af1) for the new style.

This commit implements the processing of that style: template code is
greatly simplified because the opargs no longer need to be processed as
such; expansion_context code needs to work harder to find the default
value for an oparg.

The code "works" but there appears to be a regression: the opcode VSPACE*
is no longer found. I don't think anything in this commit touches
opcodes, so it's odd. Anyway, I need to commit so I can investigate.