(WIP) Template overhaul
Template is a proper object now. No more 'details' and 'path' and 'entry'.
Usage:
  t = lbt.fn.Template.new(spec, path)
  t.register()
  ...
  t = Template.object_by_name('lbt.Questions')
  t.name   t.path   t.functions  ...

Name changes:
  arguments -> posargs
  default_options -> opargs

(kwargs is supported by Template as well, but none of the templates use
it yet, and no code does anything with it.)

Positional argument specifications are now represented internally as
simply { 1, 3 } instead of { spec = '1-3', min = 1, max = 3 }.