(WIP) CommandSpec object implemented, and command_register
It's in lbt.core because it needs to be used in both Template and
Command. It's also kind of good to emphasise such a fundamental data
type by having it in core.

Template:register() now registers all the commands of the template into
lbt.system.command_register. They are not used yet.