lbt-7-lbt-DataLoad.lua: improved 'index', implemented 'key'
So DB index is working now, thanks to the work done in the refactor.

I implemented negative indexing a la Python, so index -3 is the same as
index N-2. Any attempt to index outside the range 1..N after negatives
are accounted for results in an error.

I implemented 'key'. An attempt to index a non-existent key results in
an error.