(WIP) Various updates to new system
lbt.util.latex_expand_content_list no longer takes parameters ocr and
ol.

A couple of functions removed.

lbt.util now has `resolve_oparg` (for use in template `init` and
`expand`) and `resolve_oparg_for_macro` (for use in macros). The latter
is a name change from `get_option_for_macro` and a reimplementation
using the passed-in ExpansionContext. The former is now needed because
we no longer pass an OptionLookup object to `init` or `expand`. (Recall
that OptionLookup now only exists in the context of a Command.)

lbt.api.add_global_options bugfix.

Random, but there was a typo in Questions. The oparg MC.format was
mistyped as MC.formal.

WS0.expand(pc) updated to not take parameters ocr, ol and to use
lbt.util.resolve_oparg. Other templates will need to be updated.