(WIP) Further fixes
There was a bug resolving the oparg CHALLENGE*.color, because it should
be CHALLENGE.color. Now, when an OptionLookup object is created, if the
Command is a starred one, then we use the `command_spec.refer` value as
the opcode.

Updated a few bits of the code with new function names, etc.

In WS0, EXAMPLE and EXAMPLE* are united via the oparg 'starred'
mechanism, making the code cleaner. Ditto NOTE, CHALLENGE, HEADING.

WS1 `expand` updated like WS0 was in the previous commit.