(WIP) Fixed bug with register expansion
The Command object was being created, _then_ the registers were being
expanded into a variable that was not in the Command object. Solution:
Command:update_posargs(x) and similar for kwargs.

I considered having Command handle the register expansion, but no...
that is not its responsibility.

Perhaps ideally, we would do the register expansion before creating the
Command object. That might happen when I refactor latex_for_command.