(WIP) Reimplement latex_for_command() much more cleanly
New class LatexForCommand with method latex().

Here is the implementation of latex_for_command() in full. (Well,
excluding the supporting functions.)

  lbt.fn.latex_for_command = function (parsed_command)
    local opcode = parsed_command[1]
    local posargs = parsed_command.a
    if opcode == 'STO' then
      return impl1.handle_STO_command(posargs)
    elseif opcode == 'CTRL' then
      return impl1.handle_CTRL_command(posargs)
    else
      local lfc = lbt.fn.LatexForCommand.new(parsed_command, reg.expand_register_references)
      return lfc:latex()
    end
  end

The code is much tidier. There is more to to do tidy up lbt-3-fn.lua,
starting with removing the code that is no longer needed because of
this commit.