(WIP) All tests working again
lbt.fn.test.* gives test access to private methods.

I commented out a few assertions that rely on noX oparg parsing to be
implemented. Was it implemented before the refactor? I vaguely remember
it. Anyway, I'll need to take a look. It was on my list anyway.

(To be clear, this is the idea that `TEXT .o nopar` is treated the same
as `TEXT .o par = false`.)