Improved pragma handling
Pragmas are now based on a copy of lbt.core.DefaultPragmas. And you
access them with DEBUG instead of 'debug', etc., which gives consistency
between the LBT input and the code.

A further improvement would be to control access, so that (for instance)
calling `pragmas.FOOBAR` would raise an error. This could be achieved
with a metatable.

* Also added something about vector formatting to DEVELOPMENT.md.