BOX can now process lbt code in a .v << ... >> block
There will probably be other commands that want to do this, so perhaps I
should generalise it, but it's a useful pilot experiment for now to see
how it goes. Something to think about for the future is customising the
context in which that code is evaluated. I notice that inter-paragraph
space in a box is zero, even if it is non-zero in the surrounding
document. That is the case even without LBT evaluation, but it's worth
noting here.

Also added a note about init() to DEVELOPMENT.md. I should change it so
you can return Latex code.