Implemented \lbtDefineLatexMacros{...}
This is an improvement over the singular \lbtDefineLatexMacro because:
(a) you can define multiple macros at once;
(b) it is parsed using lbt.parser.parse_dictionary, so it is more
    flexible (allows space around =, for example)

Things to do now:
 * deprecate \lbtDefineLatexMacro
 * remove old code
 * put some new code in an impl-style namespace
 * see if error-checking needs to be improved

Also in this commit: fixed a dictionary parsing bug that I didn't know
about. Keys had to be at least two characters long!