(WIP) implementing automatic \left and \right
It's not working yet, but this is a progress commit.

* I have implemented lbt.fn.ExpansionContext.test_ctx(sources) so that
  the testing code can have an expansion context that allows macros to
  be resolved.
* simplemath:
  - smparse renamed to sm_grammar
  - oparg simplemath.leftright (defaults to true)
  - m.simplemath now takes parameter ctx instead of ignoring it (so that
    it can resolve the above oparg)
  - once the tokens are created, instead of just joining them to make a
    string (current behaviour), we process them first so that ()[] can
    be replaced with their \left and \right equivalents
     - This part does not yet work correctly, but I want to commit the
       code before working on it, as the surgery required is notable.

Fix, coming up: sm_grammar will tag each token it produces, so it will
be like { label = 'trig', value = '\\sin^2' }. In the joining process,
we will do search and replace to get \left and \right in there. However,
we must not do that search and replace on things like \text{see (1)},
which means we need to know what kind of token we are dealing with.