Added test for lbt.parser.parse_dictionary
(Pretty trivial. I was hoping to change it so a dictionary could span
multiple lines, but I'm shelving that idea for now. Actual LBT code is
very line-based. I wanted this for the settings file. That is too niche
an application to justify changing things, so the settings file will
have to be just one line for now.)