'adjustwith' is a bedrock oparg, and properly implemented
"bedrock oparg" means every command can use it, whether the command
specifies it or not. So you can have

    TEXT .o adjustwidth = 1em 3em :: Lorem ipsum...

and the margins will be brought in on the left and the right, and the
implementation of TEXT has nothing to do with it.

Two space-separated arguments are required. This uses the Latex
environment 'adjustwidth' (from the 'changepage' package), and I might
as well keep to how it's used.

I thought about implementing 'leftmargin' and 'rightmargin' instead of
'adjustwidth', but went the other way, for now.

Note: 'adjustwidth' was partially implemented before, but was not
complete, or functional.