Fixed a bug in simplemath; added debugging aid
The input "a \: b" was failing to parse because the definition of
command included \, but not \:. I added all the spacing commands I could
find to the grammar.

Moreover, we now have
 - macro 'simplemathtokens', for testing purposes, which allows you to
   see whether a string parses
 - lbt.util.simplemath_with_current_context(), which returns a function
   so that you can just call simplemath(text) and get the compiled
   result

MATH now does simplemath on its arguments using
simplemath_with_current_context (mentioned above), so it is not dropping
a \sm{...} macro into the Latex stream. After all, how do we even know
that macro has been defined?