Allow macros to run when not in an LBT expansion
A relatively recent update to macros had them receiving an expansion_id
so they could retrieve an expansion context and be sensitive to oparg
settings like vector.format.

However, introducing an expansion_id does not work too well when a
vector is called out in the wild, among plain Latex code, not in an LBT
expansion. But \V{3 -1} should be able to work in this context. It just
won't have access to any local oparg settings.

So now the system knows whether an expansion is in progress or not, and
if a macro is run with nil expansion_id then it gets access to a
skeleton context and macros can run.

* lbt.system.expansion_in_progress (true/false)
* lbt.fn.expansion_in_progress(x) -- read or write
* lbt.fn.current_expansion_id() may return nil
* lbt.fn.get_expansion_context_by_eid(eid) may return a skeleton context
* lbt.fn.ExpansionContext.skeleton_expansion_context()
* lbt.api.macro_run: comment that eid may be nil

NOTE: the code generates a skeleton context whenever needed. I could
possibly cache it.