CTRL options :: QQ.post = 3em ... CTRL options-pop :: QQ.post
You can now change options in the middle of the document. And undo the
change later.

Implementation:
 * CTRL options defers to lbt.fn.options_push, which defers to
   ExpansionContext:opargs_local_push_dictionary, which defers to
   lbt.core.DictionaryStack:push.
 * CTRL options-pop defers to lbt.fn.options_pop, which defers to
   ExpansionContext:opargs_local_pop_keys, which defers to
   lbt.core.DictionaryStack:remove_outer_key.

In both cases the oparg cache is cleared. We could be selective in
removal, but why risk the bugs for no noticeable difference in
performance?

TODO: unit test. This has been tested on a real document, but a unit
test would be more thorough.